\name{exchange_rates}
\docType{data}
\alias{exchange_rates}
\title{Returns on exchange rates of 11 currencies}
\description{
This dataset consists of 1000 daily observations on 11 currencies.
The New Zealand Dollar (NZD), the Australian Dollar (AUD), and
the Japanese Yen (JPY) - five European currencies that eventually became
part of the Euro - the Deutsch Mark (DEM), French Franc (FRF), Belgian
Franc (BEF), Netherlands Gilder (NLG) and Spanish Peso (ESP) - as well
as three additional European currencies - the British Pound(GBP), Swedish
Krona (SEK), and Swiss Franc (CHF). Returns on exchange rates are relative to the US Dollar.
}
\usage{exchange_rates}
\format{A 1000x11 Matrix}
\references{
Abel Rodriguez, Alex Lenkoski, and Adrian Dobra: Sparse
covariance estimation in heterogeneous samples. Electronical
Journal of Statistics Volume 5 (2011), 981-1014.
\url{http://projecteuclid.org/DPubS?service=UI&version=1.0&verb=Display&handle=euclid.ejs/1316092866}
}
\keyword{datasets}

