\name{plot.DPM.GGM}
\alias{plot.DPM.GGM}
\title{Generic plot function for object of class \code{DPM.GGM}}
\description{
Plots the clustering of observations using the lattice function levelplot.
}
\usage{
plot.DPM.GGM(x,...)
}
\arguments{
\item{x}{\code{DPM.GGM} object returned by \code{\link{DPM.GGM}}}
\item{...}{additional parameters passed to the plot function}
}
\seealso{
\code{\link{DPM.GGM}}
}
