\name{plot.iHMM}
\alias{plot.iHMM}
\title{Generic plot function for object of class \code{iHMM}}
\description{
Plots the clustering of observations using the lattice function levelplot.
}
\usage{
plot.iHMM(x,...)
}
\arguments{
\item{x}{\code{iHMM} object returned by \code{\link{iHMM}}}
\item{...}{additional parameters passed to the plot function}
}
\seealso{
\code{\link{iHMM}}
}
